first_django/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
pkg-resources==0.0.0
psycopg2-binary==2.8.6
Pygments==2.8.1
PySocks==1.7.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.4
